refactor: optimize
VM-178

+++ audio_to_text/main/views.py
@@ -1,3 +1,5 @@
+import uuid
+
 from django.core.files.storage import default_storage
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -12,13 +14,23 @@ class TranscribeAudio(APIView):
             return Response({"error": "No file uploaded"}, status=400)
 
         audio_file = request.FILES['file']
-        file_path = default_storage.save(f"uploads/{audio_file.name}", audio_file)
 
-        model = WhisperModel("base", device="cpu")
-        segments, _ = model.transcribe(default_storage.path(file_path))
+        # Генерация уникального имени
+        unique_name = f"{uuid.uuid4()}"
+        file_path = default_storage.save(f"uploads/{unique_name}", audio_file)
+        full_path = default_storage.path(file_path)
 
-        transcript = [segment.text.strip() for segment in segments]
+        try:
+            model = WhisperModel("base", device="cuda")
+            segments, _ = model.transcribe(full_path)
 
-        default_storage.delete(file_path)
+            transcript = [segment.text.strip() for segment in segments]
+            return Response({"lyrics": transcript})
 
-        return Response({"transcript": transcript})
+        except Exception as e:
+            return Response({"error": str(e)}, status=500)
+
+        finally:
+            # Удаляем файл независимо от результата
+            if default_storage.exists(file_path):
+                default_storage.delete(file_path)
